refactor(api): Add subpackages to RideEarlyFinishTest
