+++ backend/FleetForge/src/test/java/com/team18/FleetForge/RideEarlyFinishTest/repository/RideLocationRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.team18.FleetForge.RideEarlyFinishTest;
+package com.team18.FleetForge.RideEarlyFinishTest.repository;
 
 import static org.junit.jupiter.api.Assertions.*;
